WIP: moves all edge app related commands/files to the separate mod.
