+++ src/commands/edge_app/manifest.rs
@@ -10,7 +10,7 @@ use std::path::Path;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 
-use crate::commands::edge_app_settings::{deserialize_settings, serialize_settings, Setting};
+use crate::commands::edge_app::setting::{deserialize_settings, serialize_settings, Setting};
 use crate::commands::serde_utils::{
     deserialize_option_string_field, string_field_is_none_or_empty,
 };
@@ -332,7 +332,7 @@ pub fn beautify_error_message(error: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::commands::edge_app_settings::SettingType;
+    use crate::commands::edge_app::setting::SettingType;
     use tempfile::tempdir;
 
     fn create_test_manifest() -> EdgeAppManifest {
