+++ src/commands/edge_app/server.rs
