+++ src/commands/edge_app/setting.rs
